requirements.txt
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.10.0
defusedxml==0.7.1
filelock==3.13.4
Flask==3.0.3
fonttools==4.51.0
fsspec==2024.3.1
google==3.0.0
google-api-core==2.18.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-vision==3.7.2
googleapis-common-protos==1.63.0
grpcio==1.62.2
grpcio-status==1.62.2
httplib2==0.22.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.8.0.74
packaging==24.0
pandas==2.2.2
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==2.4.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
roboflow==1.1.7
rsa==4.9
scipy==1.13.0
seaborn==0.13.2
setuptools==69.5.1
six==1.16.0
soupsieve==2.5
supervision==0.19.0
sympy==1.12
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.0.134
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.2
